feature/ASAA-25 cleanup on map clustering
